requirements.txt
streamlit
gitpython
ollama 
